setup.nsi: Fix invalid registry path
REGPATH_WUAUPOLICY is currently set to itself and doesn't inherit REGPATH_WUPOLICY's path as a result.

Due to this, you'll end up with an error due to it being unable to set the registry key. In the case of Windows Server 2008 SP2 Enterprise at least (which is where I first encountered this), it'll result in no keys getting created for LegacyUpdate.

Tested on Windows 2000, XP, Vista, 2008, and 2008 R2. 😃