Force logging off at the end of the RunOnce process
Currently we are using EWX_FORCEIFHUNG, which seems to have problems. In
an attempt to fix this, I tried sleeping for 5 minutes, and
EWX_FORCEIFHUNG didn't take effect.

Since we're on a temporary user anyway, forcing a logoff should be safe.

Note that after this is done and the system is rebooted, only ntuser.ini
is left in the C:\Documents and Settings\LegacyUpdateTemp directory.

Fixes #113