Fix installing LegacyUpdate on Windows 8.0.
Let's do this by replacing the Windows Update Agent update with one that
is newer and which has the fix for Microsoft ADV990001 (a security
update for the Windows Update Agent that was never made available to 8.0
unless you knew to apply it manually). This works
because Windows 8 can apply updates from Windows 8 Embedded and Windows
Server 2012.

Tested with a fresh install of Windows 8 Pro retail and a complete run
of Windows Update after LegacyUpdate was installed. No ill-effects were
discovered.

Fixes #128