Install the proper Service Pack and WUA for Windows XP Embedded.
This updates the installer to install Service Pack 3 for XP Embedded,
WEPOS, and Windows Fundamentals for Legacy PCs. The standard version of
SP3 for XP is incompatible with these variants.

Note that the version of WUA used for XP Home is required for XP
Embedded and variants.

Tested with fresh installations of WEPOS and Fundamentals for Legacy
PCs, and on POSReady 2009.

Fixes #73