Add support for Windows Home Server 2011 by fixing WUA and bugs (#152)
This installs Update Rollup 4, which resolves some problems with Windows
Update failures with previous WHS updates. In addition, it fixes a data
corruption bug that should be resolved on fresh installations.

The PowerPack updates for WHS 2007 are still available and seem to not
be required for Windows Update since the WUA issues were resolved in
previous commits.

Fixes #134