Fix running LegacyUpdate's installer on CPUs prior to the Pentium 2.
This turned out to be a problem with the way NSxfer was compiled, since
it was compiled with -march=native. The version of NSxfer.dll that comes
out of this was compiled with the change documented in https://github.com/kirb/nsis-nsxfer/pull/1

Tested on Windows 2000 Pro SP4 on a Pentium MMX 233.