setup.nsi: Remove an extra debugging module from a previous commit.
This was put in place to help with determining the cause of the inverted
logic in the Windows XP activation check, but for a normal build isn't
needed anymore.

We may want to consider adding this back in in the future though, but for
now I'm removing this since it breaks the build.