Hide FirstUxWnd and enable Aero Basic
Because… reasons I wasted a few hours before I realised, this needs to be done
from a native process, and not a WOW process. (Who could have guessed
undocumented internal APIs have unexpected behavior?!) So this introduces an
x86_64 build of nsisplugin, whose sole responsibility is to be callable by
rundll32.

Also hides FirstUxWnd, added in Windows 7, which displays the “Setup is
preparing your computer for first use” message (7) or boot logo (8+).