RunOnce.nsh: Fix a bug that causes the LegacyUpdate installer to not be copied.
This causes a variety of different problems depending on what version of Windows is in use.
On Windows XP x64, Server 2008 x64, and Windows 7 32-bit (which were the three I tried
that the problem was noticable on), the lack of this file causes the kernel to do memory
corruption because the launcher for RunOnce isn't copied, even though it is added into the
Registry. On all three platforms it will routinely cause the ACPI tables to be temporarily
overwritten, as lsass crashes immediately and the kernel won't even BSOD if lsass crashes
in the way that it does. This will make affected systems fail to boot Windows until the
power cable is unplugged from the machine.

Windows 8.1 seems to be unaffected, but some security features in that version of Windows
probably prevent this from happening (or the newer kernel redesign).

At first Kirb and I thought it was a problem with the MinGW toolchain or the new version
of NSIS, but after chasing a gut feeling it turned out to be caused by commit 771c0436d

This one was pretty challenging to figure out because of the lack of diagnostics from
NSIS. The underlying cause was a register conflict.