CI: Actually use the NSIS version that we build...
There is a minor problem in both workflows where the directory that the
NSIS tarball extracts to is /opt/nsis-3.11-lfs, while the workflows are
using /opt/nsis-3.11.

This causes the CI output to use NSIS 3.09, which comes from the version
of Ubuntu that the CI system uses.

Because of this though, the illegal instruction fixes for NSIS aren't
included, and all versions generated through CI are vulnerable to
CVE-2025-43715.