Install Vista/7 service packs without extracting
I was pretty sure we did it this way because the outer SFX was failing to
pass flags down to SPInstall.exe - but it seems fine now? Even on Vista
SP1? /norestart is still broken, though.

There are too many edge cases to the SP installers, especially on Vista,
that we pretty much need to be relying on using the SP installers through
the SFX. This thing is a buggy mess.

Fixes #260, fixes #407.